refactor: 코드 모듈화
